# Back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Seguros TM",
        "description": "API para gestão de seguros, clientes, apólices e sinistros",
        "contact": {
            "email": "suporte@segurostm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/seguradora/apolices/pendentes": {
            "get": {
                "tags": [
                    "Seguradora - Apólices"
                ],
                "summary": "Listar apólices pendentes de aprovação",
                "description": "Retorna todas as apólices que aguardam aprovação da seguradora",
                "operationId": "dff629ca1ca90ae46c94aa3e8ddbdad6",
                "responses": {
                    "200": {
                        "description": "Lista de apólices pendentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_apolice": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "numero_apolice": {
                                                        "type": "string",
                                                        "example": "APO-2024-001"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pendente_aprovacao"
                                                    },
                                                    "premio_total": {
                                                        "type": "number",
                                                        "example": 1500
                                                    },
                                                    "data_inicio": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-01-15"
                                                    },
                                                    "data_fim": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-01-15"
                                                    },
                                                    "cliente": {
                                                        "type": "object"
                                                    },
                                                    "bem_segurado": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seguradora/apolices/{apolice}/aprovar": {
            "post": {
                "tags": [
                    "Seguradora - Apólices"
                ],
                "summary": "Aprovar apólice",
                "description": "Aprova uma apólice pendente, ativando-a no sistema",
                "operationId": "e3ada821f440ac1f2ceb279c5084de32",
                "parameters": [
                    {
                        "name": "apolice",
                        "in": "path",
                        "description": "ID da apólice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "observacoes": {
                                        "type": "string",
                                        "example": "Apólice aprovada conforme análise de risco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Apólice aprovada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Apólice aprovada com sucesso"
                                        },
                                        "apolice": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao aprovar apólice"
                    },
                    "403": {
                        "description": "Não autorizado - apólice não pertence à seguradora"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT. Use o token retornado no endpoint de login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}